requirements.txt
asgiref==3.11.1
Django==4.2.29,<5
et_xmlfile==2.0.0
lxml==6.0.2
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2026.1.post1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
